feat: Add image carousel functionality to starter01.html
This commit adds the necessary JavaScript code to enable an image carousel functionality to the starter01.html file. The carousel displays a sequence of images and allows the user to click on the image to trigger a function that updates the image source and animates the image to the left or right.

The carousel is implemented using an array of image file names and a focus variable to keep track of the current image being displayed. The direction variable determines whether the carousel should move to the right or left. The checkPageBounds function is used to reverse the direction when the carousel reaches the page bounds.

This feature enhances the user experience by providing an interactive element to the webpage.